feed post creation improvements
