+++ package.json
@@ -63,11 +63,11 @@
     "jest": "^29.0.3",
     "pod-install": "^0.1.0",
     "prettier": "^2.0.5",
-    "react": "19.1.8",
+    "react": "19.1.4",
     "react-native": "^0.81.6",
     "react-native-builder-bob": "^0.18.2",
     "react-native-enriched": "^0.7.0",
-    "react-test-renderer": "19.1.8",
+    "react-test-renderer": "19.1.4",
     "ts-clean": "^1.0.3",
     "ts-jest": "^29.0.1",
     "typescript": "^5.6.0",
