+++ src/components/index.ts
@@ -19,7 +19,6 @@ export * from './feed-custom-toolbar-button';
 export * from './feed-follow-pill';
 export * from './feed-new-posts-banner';
 export * from './feed-post-composer';
-export * from './feed-post-create';
 export * from './feed-post-media-attach';
 export * from './feed-post-media-gallery';
 export * from './feed-post-reactions';
